Update version information.

+++ app/src/main/java/net/dimterex/sync_client/data/entries/ConnectionsLocalModel.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "connection_settings")
 data class ConnectionsLocalModel(
-    @PrimaryKey @ColumnInfo(name = "id") val id: Long,
+    @PrimaryKey @ColumnInfo(name = "id") val id: Int,
     @ColumnInfo(name = "ip_address") var ip_address: String,
     @ColumnInfo(name = "ip_port") var ip_port: Int,
     @ColumnInfo(name = "login") var login: String,
